divar_clone/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-allauth==65.3.1
django-appconf==1.0.6
django-filter==24.3
django-imagekit==5.0.0
django-taggit==6.1.0
djangorestframework==3.15.2
Markdown==3.7
pilkit==3.0
pillow==11.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2024.2
